#168 separate views to files
